Refactored the filter + pagination support
Whilst testing, I found that the combination of filtering and pagination didn't work.
The call to `UserService.GetAll` was pre-paginating the user list results before we could filter them.
Now we can get all the users first, then filter accordingly. Depending on performance, we may need to temporarily cache the users to save repeated database hits (in quick succession).

With refactoring the filtering method, I found it tricky to debug the previous Linq/Func expression, so opted to try using the `yield returns` approach. This is totally open for discussion.

Removed the `BulkUserListItemModel` class, as we no longer explicitly need it.

Added support for a custom `PageSize` by checking for an AppSetting key, (which we need to document).